+++ flux/util.py
@@ -100,7 +100,7 @@ configs = {
 
 try:
     import ml_dtypes
-    from_torch_bf16 = lambda x: jnp.asarray(x.view(dtype=torch.uint16).numpy().view(ml_dtypes.bfloat16))
+    from_torch_bf16 = lambda x: jnp.asarray(x.view(dtype=torch.uint16).numpy().view(ml_dtypes.bfloat16)).astype(jnp.bfloat16)
 except:
     from_torch_bf16 = lambda x: jnp.asarray(x.float().numpy()).astype(jnp.bfloat16)
